Fixed some compiler warnings

+++ src/rng.cpp
@@ -40,9 +40,9 @@ SeedGenerator::next()
 {
   uint64_t cur_seed;
   cur_seed = d_seed;
-  d_seed   = getpid();
+  d_seed   = (uint64_t) getpid();
   d_seed *= 129685499;
-  d_seed += time(nullptr);
+  d_seed += (uint64_t) time(nullptr);
   d_seed *= 233755607;
   d_seed += cur_seed;
   d_seed *= 38259643;
